Move test helpers into testhelper package and re-order imports
The testing package contains only methods to create test data and
assets for tests now.
