+++ pkg/credsgen/in_memory_generator/ssh_key_test.go
@@ -1,11 +1,12 @@
 package inmemorygenerator_test
 
 import (
-	"code.cloudfoundry.org/cf-operator/pkg/credsgen"
-	inmemorygenerator "code.cloudfoundry.org/cf-operator/pkg/credsgen/in_memory_generator"
-	"code.cloudfoundry.org/cf-operator/testing"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+
+	"code.cloudfoundry.org/cf-operator/pkg/credsgen"
+	inmemorygenerator "code.cloudfoundry.org/cf-operator/pkg/credsgen/in_memory_generator"
+	helper "code.cloudfoundry.org/cf-operator/pkg/testhelper"
 )
 
 var _ = Describe("InMemoryGenerator", func() {
@@ -14,7 +15,7 @@ var _ = Describe("InMemoryGenerator", func() {
 	)
 
 	BeforeEach(func() {
-		_, log := testing.NewTestLogger()
+		_, log := helper.NewTestLogger()
 		generator = inmemorygenerator.NewInMemoryGenerator(log)
 	})
