+++ pkg/kube/controllers/extendedjob/trigger_reconciler_test.go
@@ -5,24 +5,26 @@ import (
 	"fmt"
 	"time"
 
-	. "code.cloudfoundry.org/cf-operator/pkg/kube/controllers/extendedjob"
-	cfctx "code.cloudfoundry.org/cf-operator/pkg/kube/util/context"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	"code.cloudfoundry.org/cf-operator/pkg/kube/apis/extendedjob/v1alpha1"
 	"code.cloudfoundry.org/cf-operator/pkg/kube/controllers"
+	. "code.cloudfoundry.org/cf-operator/pkg/kube/controllers/extendedjob"
 	"code.cloudfoundry.org/cf-operator/pkg/kube/controllers/fakes"
+	cfctx "code.cloudfoundry.org/cf-operator/pkg/kube/util/context"
+	helper "code.cloudfoundry.org/cf-operator/pkg/testhelper"
 	"code.cloudfoundry.org/cf-operator/testing"
+
 	"go.uber.org/zap"
 	"go.uber.org/zap/zaptest/observer"
-	"k8s.io/client-go/kubernetes/scheme"
 
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/kubernetes/scheme"
 	crc "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -88,7 +90,7 @@ var _ = Describe("TriggerReconciler", func() {
 
 		BeforeEach(func() {
 			controllers.AddToScheme(scheme.Scheme)
-			logs, log = testing.NewTestLogger()
+			logs, log = helper.NewTestLogger()
 			ctrsConfig = &cfctx.Config{ //Set the context to be TODO
 				CtxTimeOut: 10 * time.Second,
 				CtxType:    cfctx.NewContext(),
