+++ pkg/testhelper/logger.go
@@ -1,4 +1,4 @@
-package testing
+package testhelper
 
 import (
 	"fmt"
