+++ pkg/testhelper/randomstring.go
@@ -1,4 +1,4 @@
-package testing
+package testhelper
 
 import "math/rand"
